move: bid 엔티티 디렉토리 변경
