+++ auction-service/src/main/java/com/javaauction/auction_service/domain/entity/enums/BidStatus.java
@@ -1,4 +1,4 @@
-package com.javaauction.auction_service.domain.model.enums;
+package com.javaauction.auction_service.domain.entity.enums;
 
 public enum BidStatus {
     HELD,
