+++ auction-service/src/main/java/com/javaauction/auction_service/presentation/dto/response/ResPostBidDto.java
@@ -1,6 +1,6 @@
 package com.javaauction.auction_service.presentation.dto.response;
 
-import com.javaauction.auction_service.domain.model.enums.BidStatus;
+import com.javaauction.auction_service.domain.entity.enums.BidStatus;
 import lombok.AllArgsConstructor; import lombok.Builder;
 import lombok.Getter;
